+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Changed
+- Reduced number of external dependencies by removing commons-lang3, commons-io, commons-codec.
+
 ## [2.1.4] - 2022-01-14
 
 ### Fixed